Migrate src/components/Inputs/StringField
