+++ src/main/webapp/ui/src/components/Inputs/StringField.js.flow
@@ -1,0 +1,33 @@
+// @flow
+
+import React, { type Node } from "react";
+import { type Sx } from "../../util/styles";
+
+export type StringFieldArgs = {|
+  // required
+  value: string,
+
+  // optional
+  autoFocus?: boolean,
+  disabled?: boolean,
+  error?: boolean,
+  InputProps?: {|
+    sx?: Sx,
+    startAdornment?: Node,
+    endAdornment?: Node,
+  |},
+  name?: string,
+  noValueLabel?: ?string,
+  onBlur?: () => void,
+  onChange?: ({| target: { value: string } |}) => void,
+  onFocus?: ({| target: HTMLInputElement |}) => void,
+  onKeyDown?: (KeyboardEvent) => void,
+  size?: "small" | "medium",
+  variant?: "filled" | "outlined" | "standard",
+  minLength?: number,
+  "data-testid"?: string,
+  fullWidth?: boolean,
+  id?: string,
+|};
+
+declare export default function StringField(StringFieldArgs): Node;
